Make Timepicker readonly, refactor 'clearable' method
